+++ src/DiscordBot/deploy-commands.ts
@@ -1,6 +1,6 @@
 import { REST } from "@discordjs/rest";
 import { Routes } from "discord-api-types/v9";
-import { clientId, guildId, token } from "../../profile.json";
+import { clientId, guildId, token } from "profile.json";
 import fs from "fs";
 
 export async function deployCommands() {
