+++ src/main.ts
@@ -1,4 +1,4 @@
-import ui from "./src/UI/ui";
+import ui from "UI/ui";
 
 const arg = process.argv.pop();
 ui.init(arg);
