+++ tsconfig.json
@@ -7,6 +7,17 @@
     "target": "es5",                                /* Specify ECMAScript target version: 'ES3' (default), 'ES5', 'ES2015', 'ES2016', 'ES2017', 'ES2018', 'ES2019', 'ES2020', 'ES2021', or 'ESNEXT'. */
     "module": "commonjs",  
     "resolveJsonModule": true,
+    "baseUrl": "./src",
+    "paths": {
+      "ani/*" : ["Anilist/*"],
+      "db/*" : ["database/*"],
+      "discord/*" : ["DiscordBot/*"],
+      "nyaa/*" : ["Nyaa/*"],
+      "qbit/*" : ["qBitTorrent/*"],
+      "scheduler/*" : ["Scheduler/*"],
+      "ui/*" : ["UI/*"],
+      "utils/*" : ["utils/*"],
+    },
     // "lib": [],                                   /* Specify library files to be included in the compilation. */
     // "allowJs": true,                             /* Allow javascript files to be compiled. */
     // "checkJs": true,                             /* Report errors in .js files. */
